Changes by tilda

+++ src/main/java/com/programvaruprojekt/springbatchtutorial/model/Transaction.java
@@ -36,11 +36,11 @@ public class Transaction {
     }
 
     // Getters and setters
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
